+++ Android/app/src/main/java/com/felipearpa/tyche/network/LocalUrlBasePathProvider.kt
@@ -4,5 +4,5 @@ import com.felipearpa.tyche.BuildConfig
 import com.felipearpa.tyche.core.network.UrlBasePathProvider
 
 class LocalUrlBasePathProvider : UrlBasePathProvider {
-    override val basePath: String = BuildConfig.URL_BASE_BATH
+    override val basePath: String = BuildConfig.URL_BASE_PATH
 }