+++ Android/data/pool/src/main/java/com/felipearpa/tyche/data/pool/domain/PoolGamblerScoreResponseTransformer.kt
@@ -2,11 +2,12 @@ package com.felipearpa.tyche.data.pool.domain
 
 internal fun PoolGamblerScoreResponse.toPoolGamblerScore() =
     PoolGamblerScore(
-        poolId = this.poolId,
-        poolName = this.poolName,
-        gamblerId = this.gamblerId,
-        gamblerUsername = this.gamblerUsername,
-        position = this.position,
-        beforePosition = this.beforePosition,
-        score = this.score,
+        poolId = poolId,
+        poolName = poolName,
+        gamblerId = gamblerId,
+        gamblerUsername = gamblerUsername,
+        position = position,
+        beforePosition = beforePosition,
+        score = score,
+        gamblerCount = gamblerCount,
     )